% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_pos_rel.R
\name{calc_element_margin_data}
\alias{calc_element_margin_data}
\title{Title}
\usage{
calc_element_margin_data(.plot_obj)
}
\arguments{
\item{.plot_obj}{...}
}
\value{
tibble()
}
\description{
Title
}
\examples{
NULL
}
