% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottr-package.R
\docType{package}
\name{plottr-package}
\alias{plottr}
\alias{plottr-package}
\title{plottr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
