% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_eccdf.R
\name{stat_eccdf}
\alias{stat_eccdf}
\title{stat_eccdf}
\usage{
stat_eccdf(
  mapping = NULL,
  data = NULL,
  geom = "step",
  position = "identity",
  ...,
  n = NULL,
  pad = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{...}

\item{data}{...}

\item{geom}{...}

\item{position}{...}

\item{...}{...}

\item{n}{...}

\item{pad}{...}

\item{na.rm}{...}

\item{show.legend}{...}

\item{inherit.aes}{...}
}
\value{
...
}
\description{
stat_eccdf
}
\examples{
NULL
}
