% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_to_file}
\alias{plot_to_file}
\title{try to install "preview" when using tinytex distro}
\usage{
plot_to_file(
  .plot,
  .plot_name = NULL,
  .figure_dir = character(0L),
  .width = NULL,
  .height = NULL,
  .units = "mm",
  .png_dpi = 600,
  .knit = TRUE,
  .fonts = list(main = "TeX Gyre Termes", math = NULL, mono = NULL)
)
}
\arguments{
\item{.plot}{A number.}

\item{.plot_name}{...}

\item{.figure_dir}{...}

\item{.width}{...}

\item{.height}{...}

\item{.units}{...}

\item{.png_dpi}{...}

\item{.knit}{...}

\item{.fonts}{...}
}
\value{
If .knit ...
}
\description{
try to install "preview" when using tinytex distro
}
\examples{
# plot_to_file(
#   ggplot2::ggplot(),
#   "test_plot",
#   .figure_dir=tempdir(),
#   .width=10,
#   .height=10,
#   .units="cm"
# )
}
